# src/swagger/swaggerApi2.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Movie",
    "version": "2.0.0",
    "description": "API RESTful pour la gestion des utilisateurs, des films  et l’authentification sécurisée"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api/v2"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Lister les utilisateurs",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs"
          }
        }
      },
      "post": {
        "summary": "Créer un nouvel utilisateur",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé"
          },
          "400": {
            "description": "Erreur de requête"
          }
        }
      }
    },

    "/login": {
      "post": {
        "summary": "Authentifier un utilisateur",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie"
          },
          "401": {
            "description": "Échec de l'authentification"
          }
        }
      }
    },

    "/movie": {
      "get": {
        "summary": "Lister les films",
        "tags": ["Movies"],
        "responses": {
          "200": {
            "description": "Liste des films"
          },
          "401": {
            "description": "Non autorisé"
          }
        }
      },
      "post": {
        "summary": "Créer un film",
        "tags": ["Movies"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Film créé"
          },
          "400": {
            "description": "Requête invalide"
          }
        }
      }
    }
  },

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserCreate": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "alice42" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 }
        },
        "required": ["username", "email", "password"]
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "MovieCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Inception" },
          "genre": { "type": "string", "example": "Sci-Fi" },
          "durationMin": { "type": "integer", "example": 148 },
          "synopsis": { "type": "string", "example": "A thief who steals corporate secrets through dream-sharing..." }
        },
        "required": ["title", "genre", "durationMin", "synopsis"]
      }
    }
  }
}